Improve testability and structure of APIClient
