+++ Ptt/API/Models/APIModel.swift
@@ -1,0 +1,10 @@
+//
+//  APIModel.swift
+//  Ptt
+//
+//  Created by Anson on 2020/11/9.
+//  Copyright © 2020 Ptt. All rights reserved.
+//
+
+import Foundation
+struct APIModel {}
